# server/docs/swagger.admin.yaml
openapi: 3.1.0
info:
  title: Admin Catalog API
  version: 1.0.0
  summary: API админки для управления группами товаров и офферами
  description: >
    # Подробная документация для фронтенда


    Данный API предназначен для управления каталогом в админке.


    Архитектура каталога разделена на 2 сущности:


    1. **Группа товара** (`ProductGroup`)   Это общая сущность товара:   -
    slug   - title   - subtitle   - description   - imageURL
       - categoryIds
       - status
       - variationAxes
       - characteristics

    2. **Оффер** (`Offer`)   Это конкретная вариация товара внутри группы:   -
    sku   - price   - opt_price   - available   - img
       - optionMap
       - optionValues
       - optionKey
       - characteristics
       - stocks

    ---


    ## Важнейшая концепция


    ### variationAxes

    На уровне группы задаются оси вариаций.


    Пример:

    - `A1` = вкус

    - `A2` = количество

    - `A3` = формат упаковки


    ### optionMap

    У каждого оффера есть `optionMap`, который хранит выбранные значения по всем
    осям.


    Пример:

    ```json

    {  "A1": "berry",  "A2": 60,  "A3": "box"

    }

    ```


    ### Что фронт НЕ должен сам вычислять

    Бэкенд автоматически строит:

    - `optionValues`

    - `optionKey`


    Поэтому при создании/обновлении фронт **не должен** присылать:

    - `optionValues`

    - `optionKey`


    Фронт присылает:

    - `optionMap`

    - остальные поля оффера


    ---


...


...

...


...


...

...

...

...
    endpoints
...
servers:
...
paths:
...
  /catalog/admin/catalog/groups/{groupId}:
    get:
      tags:
        - Admin Catalog Groups
      summary: Получить одну группу товара для админки
      description: >
        Возвращает группу товара для экрана редактирования.


        Офферы не включаются в ответ.


        Это сделано специально для производительности.

        Если у товара 400+ офферов, то тащить их всегда в одном ответе — плохая
        идея.


        1. Получить группу через этот endpoint

        2. Затем отдельно получить офферы через:   `GET
        /catalog/admin/catalog/groups/{groupId}/offers`


        - маленький товар

        - debug режим

        - спец-экран полной пересборки
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
        - name: includeOffers
          in: query
          required: false
          description: Включить ли офферы в этот же ответ
          schema:
            type: boolean
            default: false
        - name: offersPage
          in: query
          required: false
          description: Страница офферов, если includeOffers=true
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: offersLimit
          in: query
          required: false
          description: Лимит офферов, если includeOffers=true
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Группа успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGroupGetResponse'
              examples:
                withoutOffers:
                  summary: Группа без офферов
                  value:
                    item:
                      _id: 69aaf53efcc1f31d1fda65ed
                      slug: vitamin-c-liposomal-t003
                      title:
                        ua: Вітамін C Ліпосомальний (t003)
                        en: Витамин C Липосомальный (t003)
                      subtitle:
                        ua: Ніжний догляд для щоденного використання
                        en: Нежный уход для ежедневного использования
                      description:
                        ua: Якісний товар для щоденного використання.
                        en: Качественный товар для ежедневного использования.
                      categoryIds:
                        - 696f9c4ab7919bf78623d905
                      imageURL: https://i.postimg.cc/8k2LRmzP/fallback.webp
                      variationAxes:
                        - axisId: A1
                          title:
                            ua: Смак
                            en: Вкус
                          type: select
                          unit: null
                          valuesPreset:
                            - orange
                            - berry
                      characteristics: []
                      status: active
                      createdAt: '2026-03-06T15:39:42.311Z'
                      updatedAt: '2026-03-27T09:31:21.114Z'
                      offers: []
                      offersMeta:
                        page: 1
                        limit: 50
                        total: 0
                        pages: 0
                        included: false
                withOffers:
                  summary: Группа с вложенной страницей офферов
                  value:
                    item:
                      _id: 69aaf53efcc1f31d1fda65ed
                      slug: vitamin-c-liposomal-t003
                      title:
                        ua: Вітамін C Ліпосомальний (t003)
                        en: Витамин C Липосомальный (t003)
                      subtitle:
                        ua: Ніжний догляд для щоденного використання
                        en: Нежный уход для ежедневного использования
                      description:
                        ua: Якісний товар для щоденного використання.
                        en: Качественный товар для ежедневного использования.
                      categoryIds:
                        - 696f9c4ab7919bf78623d905
                      imageURL: https://i.postimg.cc/8k2LRmzP/fallback.webp
                      variationAxes:
                        - axisId: A1
                          title:
                            ua: Смак
                            en: Вкус
                          type: select
                          unit: null
                          valuesPreset:
                            - orange
                            - berry
                        - axisId: A2
                          title:
                            ua: Кількість
                            en: Количество
                          type: number
                          unit: caps
                          valuesPreset:
                            - 30
                            - 60
                      characteristics: []
                      status: active
                      createdAt: '2026-03-06T15:39:42.311Z'
                      updatedAt: '2026-03-27T09:31:21.114Z'
                      offers:
                        - _id: 69aaf53efcc1f31d1fda65f4
                          groupId: 69aaf53efcc1f31d1fda65ed
                          sku: VITC-ORANGE-30-t003
                          price: 593
                          opt_price: null
                          available: true
                          img: https://i.postimg.cc/8k2LRmzP/fallback.webp
                          optionMap:
                            A1: orange
                            A2: 30
                          optionValues:
                            - orange
                            - 30
                          optionKey: orange|30
                          stocks:
                            - warehouseId: 69c470e7475b219c3e3255a0
                              onHand: 10
                              reserved: 0
                          characteristics: []
                          createdAt: '2026-03-06T15:39:42.311Z'
                          updatedAt: '2026-03-25T23:34:00.500Z'
                      offersMeta:
                        page: 1
                        limit: 50
                        total: 4
                        pages: 1
                        included: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Admin Catalog Groups
      summary: Полностью обновить группу и матрицу офферов
      description: >
        Тяжелая операция полного обновления.


        - изменились variationAxes

        - пересобирается вся матрица офферов

        - массовый импорт / bulk save

        - нужно заменить полный набор офферов


        Этот endpoint ожидает **полный актуальный снимок состояния**:

        - поля группы

        - все офферы, которые должны остаться


        Офферы, которых нет в payload, могут быть удалены.


        Не надо использовать этот endpoint для:

        - изменения price у одного оффера

        - изменения available у одного оффера

        - изменения stocks у одного оффера


        Для этого есть PATCH endpoints.


        Тогда все офферы должны соответствовать новым осям.

        Например, если добавлена новая ось `A3`, каждый оффер должен содержать
        `optionMap.A3`.
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGroupRequest'
            examples:
              rebuildAll:
                summary: Полная пересборка группы и офферов
                value:
                  slug: vitamin-c-liposomal-t003
                  categoryIds:
                    - 696f9c4ab7919bf78623d905
                  status: active
                  imageURL: https://i.postimg.cc/8k2LRmzP/fallback.webp
                  title:
                    ua: Вітамін C Ліпосомальний (t003)
                    en: Витамин C Липосомальный (t003)
                  subtitle:
                    ua: Ніжний догляд для щоденного використання
                    en: Нежный уход для ежедневного использования
                  description:
                    ua: Оновлений опис товару.
                    en: Обновленное описание товара.
                  variationAxes:
                    - axisId: A1
                      title:
                        ua: Смак
                        en: Вкус
                      type: select
                      unit: null
                      valuesPreset:
                        - orange
                        - berry
                    - axisId: A2
                      title:
                        ua: Кількість
                        en: Количество
                      type: number
                      unit: caps
                      valuesPreset:
                        - 30
                        - 60
                    - axisId: A3
                      title:
                        ua: Формат
                        en: Формат
                      type: select
                      unit: null
                      valuesPreset:
                        - box
                        - pouch
                  characteristics:
                    - key: brand
                      type: select
                      unit: null
                      value:
                        value: MST
                        label:
                          ua: MST
                          en: MST
                      values: []
                  offers:
                    - _id: 69aaf53efcc1f31d1fda65f4
                      sku: VITC-ORANGE-30-BOX-t003
                      price: 593
                      opt_price: null
                      available: true
                      img: https://i.postimg.cc/8k2LRmzP/fallback.webp
                      optionMap:
                        A1: orange
                        A2: 30
                        A3: box
                      stocks:
                        - warehouseId: 69c470e7475b219c3e3255a0
                          onHand: 10
                          reserved: 0
                      characteristics: []
                    - _id: 69aaf53efcc1f31d1fda65f5
                      sku: VITC-ORANGE-60-BOX-t003
                      price: 766
                      opt_price: null
                      available: true
                      img: https://i.postimg.cc/8k2LRmzP/fallback.webp
                      optionMap:
                        A1: orange
                        A2: 60
                        A3: box
                      stocks:
                        - warehouseId: 69c470e7475b219c3e3255a0
                          onHand: 5
                          reserved: 0
                      characteristics: []
                    - sku: VITC-BERRY-60-POUCH-t003
                      price: 1299
                      opt_price: null
                      available: true
                      img: https://i.postimg.cc/8k2LRmzP/fallback.webp
                      optionMap:
                        A1: berry
                        A2: 60
                        A3: pouch
                      stocks:
                        - warehouseId: 69c470e7475b219c3e3255a0
                          onHand: 8
                          reserved: 0
                      characteristics:
                        - key: tags
                          type: multiselect
                          unit: null
                          value: null
                          values:
                            - value: new
                              label:
                                ua: Новинка
                                en: Новинка
      responses:
        '200':
          description: Группа успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGroupMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    patch:
      tags:
        - Admin Catalog Groups
      summary: Частично обновить только поля группы
      description: |
        Частично обновляет только поля группы.

        - slug
        - title
        - subtitle
        - description
        - status
        - imageURL
        - categoryIds
        - characteristics

        - variationAxes
        - offers

        Если нужно менять variationAxes или набор офферов, используйте:
        - `PUT /catalog/admin/catalog/groups/{groupId}`
        - или offer endpoints

        Использовать для обычной формы группы в верхней части страницы.
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchGroupRequest'
            examples:
              patchBasic:
                summary: Обновить название и статус
                value:
                  title:
                    ua: Оновлена назва
                    en: Обновленное название
                  subtitle:
                    ua: Оновлений підзаголовок
                    en: Обновленный подзаголовок
                  status: active
      responses:
        '200':
          description: Группа успешно обновлена частично
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminGroupMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags:
        - Admin Catalog Groups
      summary: Удалить группу и все её офферы
      description: >
        Удаляет группу и все связанные офферы.


        Это hard delete.


        Перед удалением показывать подтверждение.

        Во многих бизнес-сценариях безопаснее использовать статус `hidden`, чем
        физическое удаление.
      parameters:
        - $ref: '#/components/parameters/GroupIdPath'
      responses:
        '200':
          description: Группа успешно удалена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
              examples:
                success:
                  value:
                    ok: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    GroupIdPath:
      name: groupId
      in: path
      required: true
      description: Mongo ObjectId группы товара
      schema:
        $ref: '#/components/schemas/ObjectIdString'
...
  responses:
    BadRequest:
      description: Ошибка валидации запроса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidGroupId:
              value:
                message: Invalid groupId
                code: BAD_REQUEST
            invalidOfferId:
              value:
                message: Invalid offerId
                code: BAD_REQUEST
            invalidPayload:
              value:
                message: >-
                  offers[0].optionMap.A3 is required because this axis exists in
                  variationAxes
                code: BAD_REQUEST
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            groupNotFound:
              value:
                message: ProductGroup not found
                code: NOT_FOUND
            offerNotFound:
              value:
                message: Offer not found
                code: NOT_FOUND
    Conflict:
      description: Конфликт данных
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            duplicateSlug:
              value:
                message: Slug already exists
                code: CONFLICT
            duplicateSku:
              value:
                message: SKU already exists - VITC-BERRY-60-BOX-t003
                code: CONFLICT
            duplicateVariant:
              value:
                message: Variant combination already exists - berry|60|box
                code: CONFLICT
  schemas:
    ObjectIdString:
      type: string
      pattern: ^[a-fA-F0-9]{24}$
      examples:
        - 69aaf53efcc1f31d1fda65ed
...
    GroupStatus:
      type: string
      enum:
        - active
        - draft
        - hidden
    LocalizedText:
      type: object
      additionalProperties: false
      required:
        - ua
        - en
      properties:
        ua:
          type: string
        en:
          type: string
      example:
        ua: Вітамін C
        en: Витамин C
...
    CharacteristicType:
      type: string
      enum:
        - string
        - number
        - boolean
        - select
        - multiselect
    CharacteristicInput:
      type: object
      additionalProperties: false
      required:
        - key
        - type
      properties:
        key:
          type: string
          example: brand
        type:
          $ref: '#/components/schemas/CharacteristicType'
        unit:
          oneOf:
            - type: string
            - type: 'null'
          nullable: true
          example: caps
        value:
          nullable: true
          description: |
            Универсальное поле значения.

            Возможные реальные форматы:
            - строка
            - число
            - boolean
            - объект вида `{ value, label }`
            - локализованный объект, если бизнес-логика это допускает
        values:
          type: array
          items:
            nullable: true
          description: |
            Используется чаще всего для multiselect.
            Для остальных типов обычно пустой массив.
      description: |
        Универсальная структура характеристики.
        Используется и для группы, и для оффера.

        Примеры:

        ```json
        {
          "key": "brand",
          "type": "select",
          "unit": null,
          "value": {
            "value": "MST",
            "label": {
              "ua": "MST",
              "en": "MST"
            }
          },
          "values": []
        }
        ```

        ```json
        {
          "key": "tags",
          "type": "multiselect",
          "unit": null,
          "value": null,
          "values": [
            {
              "value": "new",
              "label": {
                "ua": "Новинка",
                "en": "Новинка"
              }
            }
          ]
        }
        ```
    VariationAxisInput:
      type: object
      additionalProperties: false
      required:
        - axisId
        - title
        - type
        - valuesPreset
      properties:
        axisId:
          type: string
          description: Стабильный идентификатор оси вариации
          example: A1
        title:
          $ref: '#/components/schemas/LocalizedText'
        type:
          type: string
          enum:
            - number
            - select
            - string
        unit:
          oneOf:
            - type: string
            - type: 'null'
          nullable: true
        valuesPreset:
          type: array
          items:
            nullable: true
      description: |
        Ось вариации на уровне группы.

        Пример:
        ```json
        {
          "axisId": "A2",
          "title": {
            "ua": "Кількість",
            "en": "Количество"
          },
          "type": "number",
          "unit": "caps",
          "valuesPreset": [30, 60]
        }
        ```
    StockRowInput:
      type: object
      additionalProperties: false
      required:
        - warehouseId
        - onHand
        - reserved
      properties:
        warehouseId:
          $ref: '#/components/schemas/ObjectIdString'
        onHand:
          type: number
        reserved:
          type: number
      example:
        warehouseId: 69c470e7475b219c3e3255a0
        onHand: 10
        reserved: 0
    OfferOptionMap:
      type: object
      additionalProperties: true
      description: >
        Объект соответствия `axisId -> значение`.


        Обязательно должен содержать значения для всех текущих variationAxes
        группы.


        Пример:

        ```json

        {  "A1": "orange",  "A2": 30,  "A3": "box"

        }

        ```
    CreateOfferRequest:
      type: object
      additionalProperties: false
      required:
        - sku
        - price
        - available
        - img
        - optionMap
        - stocks
        - characteristics
      properties:
        sku:
          type: string
          example: VITC-ORANGE-30-t003
        price:
          type: number
          minimum: 0
          example: 593
        opt_price:
          oneOf:
            - type: number
              minimum: 0
            - type: 'null'
          nullable: true
          example: null
        available:
          type: boolean
          example: true
        img:
          type: string
          example: https://i.postimg.cc/8k2LRmzP/fallback.webp
        optionMap:
          $ref: '#/components/schemas/OfferOptionMap'
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockRowInput'
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicInput'
      description: |
        Payload для создания оффера.

        Очень важно:
        - optionValues не передавать
        - optionKey не передавать
...
    CreateGroupRequest:
      type: object
      additionalProperties: false
      required:
        - slug
        - categoryIds
        - status
        - imageURL
        - title
        - subtitle
        - description
        - variationAxes
        - characteristics
        - offers
      properties:
        slug:
          type: string
          example: vitamin-c-liposomal-t003
        categoryIds:
          type: array
          items:
            $ref: '#/components/schemas/ObjectIdString'
        status:
          $ref: '#/components/schemas/GroupStatus'
        imageURL:
          type: string
        title:
          $ref: '#/components/schemas/LocalizedText'
        subtitle:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        variationAxes:
          type: array
          items:
            $ref: '#/components/schemas/VariationAxisInput'
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicInput'
        offers:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/CreateOfferRequest'
              - allOf:
                  - $ref: '#/components/schemas/CreateOfferRequest'
                  - type: object
                    properties:
                      _id:
                        $ref: '#/components/schemas/ObjectIdString'
    UpdateGroupRequest:
      allOf:
        - $ref: '#/components/schemas/CreateGroupRequest'
      description: |
        Полный снимок состояния группы и всех офферов.

        Важно:
        - при update фронт должен отправлять все офферы, которые должны остаться
        - отсутствующие офферы могут быть удалены
        - если variationAxes изменились, все optionMap должны быть актуальны
    PatchGroupRequest:
      type: object
      additionalProperties: false
      properties:
        slug:
          type: string
        title:
          $ref: '#/components/schemas/LocalizedText'
        subtitle:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        status:
          $ref: '#/components/schemas/GroupStatus'
        imageURL:
          type: string
        categoryIds:
          type: array
          items:
            $ref: '#/components/schemas/ObjectIdString'
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicInput'
      description: |
        Частичный payload только для полей группы.
        Нельзя передавать variationAxes и offers.
    OfferOutput:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectIdString'
        groupId:
          $ref: '#/components/schemas/ObjectIdString'
        sku:
          type: string
        price:
          type: number
        opt_price:
          oneOf:
            - type: number
            - type: 'null'
          nullable: true
        available:
          type: boolean
        img:
          type: string
        optionMap:
          type: object
          additionalProperties: true
        optionValues:
          type: array
          items:
            nullable: true
        optionKey:
          type: string
        stocks:
          type: array
          items:
            $ref: '#/components/schemas/StockRowInput'
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicInput'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GroupOutput:
      type: object
      properties:
        _id:
          $ref: '#/components/schemas/ObjectIdString'
        slug:
          type: string
        title:
          $ref: '#/components/schemas/LocalizedText'
        subtitle:
          $ref: '#/components/schemas/LocalizedText'
        description:
          $ref: '#/components/schemas/LocalizedText'
        categoryIds:
          type: array
          items:
            $ref: '#/components/schemas/ObjectIdString'
        imageURL:
          type: string
        variationAxes:
          type: array
          items:
            $ref: '#/components/schemas/VariationAxisInput'
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/CharacteristicInput'
        status:
          $ref: '#/components/schemas/GroupStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - pages
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    OffersMeta:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required:
            - included
          properties:
            included:
              type: boolean
...
    AdminGroupGetResponse:
      type: object
      required:
        - item
      properties:
        item:
          allOf:
            - $ref: '#/components/schemas/GroupOutput'
            - type: object
              properties:
                offers:
                  type: array
                  items:
                    $ref: '#/components/schemas/OfferOutput'
                offersMeta:
                  $ref: '#/components/schemas/OffersMeta'
    AdminGroupMutationResponse:
      type: object
      required:
        - ok
        - item
      properties:
        ok:
          type: boolean
        item:
          allOf:
            - $ref: '#/components/schemas/GroupOutput'
            - type: object
              properties:
                offersMeta:
                  $ref: '#/components/schemas/OffersMeta'
...
    OkResponse:
      type: object
      required:
        - ok
      properties:
        ok:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        code:
          type: string
        details:
          nullable: true
...